+++ src/ta_func/ta_LINEARREG_INTERCEPT.c
@@ -138,8 +138,8 @@ TA_LIB_API TA_RetCode TA_LINEARREG_INTERCEPT( int    startIdx,
    /* Index into the output. */
    today = startIdx;
    trailingIdx = startIdx - lookbackTotal;
-   SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-   SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+   SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+   SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
    Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
    /* Prime the two data-dependent window sums for the first output with a
     * one-time full-window scan. SumX/SumXSqr/Divisor are period-only constants;
@@ -217,8 +217,8 @@ TA_LIB_API TA_RetCode TA_LINEARREG_INTERCEPT_Unguarded( int    startIdx,
    outIdx = 0;
    today = startIdx;
    trailingIdx = startIdx - lookbackTotal;
-   SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-   SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+   SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+   SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
    Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
    SumXY = 0;
    SumY = 0;
@@ -296,8 +296,8 @@ TA_RetCode TA_S_LINEARREG_INTERCEPT( int    startIdx,
    outIdx = 0;
    today = startIdx;
    trailingIdx = startIdx - lookbackTotal;
-   SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-   SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+   SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+   SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
    Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
    SumXY = 0;
    SumY = 0;
@@ -361,8 +361,8 @@ TA_RetCode TA_S_LINEARREG_INTERCEPT_Unguarded( int    startIdx,
    outIdx = 0;
    today = startIdx;
    trailingIdx = startIdx - lookbackTotal;
-   SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-   SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+   SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+   SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
    Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
    SumXY = 0;
    SumY = 0;
@@ -507,8 +507,8 @@ TA_RetCode TA_LINEARREG_INTERCEPT_OpenInternal( struct TA_LINEARREG_INTERCEPT_St
       /* Index into the output. */
       today = startIdx;
       trailingIdx = startIdx - lookbackTotal;
-      SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-      SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+      SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+      SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
       Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
       /* Prime the two data-dependent window sums for the first output with a
        * one-time full-window scan. SumX/SumXSqr/Divisor are period-only constants;
@@ -650,8 +650,8 @@ TA_LIB_API TA_RetCode TA_LINEARREG_INTERCEPT_OpenAndFill( TA_LINEARREG_INTERCEPT
       /* Index into the output. */
       today = startIdx;
       trailingIdx = startIdx - lookbackTotal;
-      SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-      SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+      SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+      SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
       Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
       /* Prime the two data-dependent window sums for the first output with a
        * one-time full-window scan. SumX/SumXSqr/Divisor are period-only constants;
