+++ ta_codegen/output/java/library/fragments/Core_TSF.java
@@ -90,8 +90,8 @@
       /* Index into the output. */
       today = startIdx;
       trailingIdx = startIdx - lookbackTotal;
-      SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-      SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+      SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+      SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
       Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
       /* Prime the two data-dependent window sums for the first output with a
        * one-time full-window scan. SumX/SumXSqr/Divisor are period-only constants;
@@ -166,8 +166,8 @@
       outIdx = 0;
       today = startIdx;
       trailingIdx = startIdx - lookbackTotal;
-      SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-      SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+      SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+      SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
       Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
       SumXY = 0;
       SumY = 0;
@@ -239,8 +239,8 @@
       outIdx = 0;
       today = startIdx;
       trailingIdx = startIdx - lookbackTotal;
-      SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-      SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+      SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+      SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
       Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
       SumXY = 0;
       SumY = 0;
@@ -301,8 +301,8 @@
       outIdx = 0;
       today = startIdx;
       trailingIdx = startIdx - lookbackTotal;
-      SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-      SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+      SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+      SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
       Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
       SumXY = 0;
       SumY = 0;
@@ -493,8 +493,8 @@
       /* Index into the output. */
       today = startIdx;
       trailingIdx = startIdx - lookbackTotal;
-      SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-      SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+      SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+      SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
       Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
       /* Prime the two data-dependent window sums for the first output with a
        * one-time full-window scan. SumX/SumXSqr/Divisor are period-only constants;
@@ -614,8 +614,8 @@
       /* Index into the output. */
       today = startIdx;
       trailingIdx = startIdx - lookbackTotal;
-      SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-      SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+      SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+      SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
       Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
       /* Prime the two data-dependent window sums for the first output with a
        * one-time full-window scan. SumX/SumXSqr/Divisor are period-only constants;
